pyproject.toml
[tool.poetry]
name = "skelshop"
version = "0.1.0"
description = "Utilities for dealing with pose skeleton data, currently from OpenPose"
authors = ["Frankie Robertson <frankie@robertson.name>"]
license = "MIT"

[tool.poetry.scripts]
skelshop = 'skelshop.cmd.skelshop:skelshop'

[tool.poetry.dependencies]
python = "^3.7"
h5py = "^3.1.0"
imutils = "^0.5.3"
click = "^7.1.2"
opencv-python-headless = "^4.2.0"
torch = "^1.6.0"
more_itertools = "^8.3.0"
click_log = "^0.3.2"
ordered-set = "^4.0.1"
orjson = "^3.1.0"
numpy = "^1.19.0"
decord = "==0.4.2"

snakemake = { version = "^5.18.1", optional = true}
scenedetect = { version = "^0.5.2", optional = true}
pygame = { version = "^2.0.0", optional = true }
numba = { version = ">=0.49.1", optional = true }
matplotlib = { version = "^3.2.1", optional = true }
scikit-learn = { git = "https://github.com/frankier/scikit-learn", rev = "9fcdf315536eb6c40257a082d0f72fe74b2fed08", optional = true }
scipy = { version = "^1.5.0", optional = true }
face_recognition = { version = "^1.3.0", optional = true }
pandas = { version = "^1.1.4", optional = true }
seaborn = { version = "^0.11.0", optional = true }
pyarrow = { version = "^2.0.0", optional = true }
dlib = {git = "https://github.com/frankier/dlib", rev = "get_face_chip_details-python", optional = true}
imagesize = "^1.2.0"
hdf5plugin = "^2.3.1"
faiss-cpu = { version = "^1.6.5", optional = true }
SPARQLWrapper = { version = "^1.8.5", optional = true }
certifi = { version = "^2020.12.5", optional = true }
ray = {version = "^1.0.1", optional = true}
joblib = {version = "^1.0.0", optional = true}
sklearn-ann = {git = "https://github.com/frankier/sklearn-ann.git", rev = "7e0a0cf852240a21b0c5fd23d8ddf0c8b85d7d8e"}
pynndescent = {version = "^0.5.1", optional = true}
tqdm = "^4.52.0"

[tool.poetry.extras]
pipeline = ["scenedetect", "snakemake"]
play = ["pygame"]
face = ["face_recognition", "dlib"]
ssmat = ["numba", "matplotlib", "scikit-learn", "scipy"]
calibrate = ["pandas", "seaborn", "pyarrow"]
buildrefs = ["SPARQLWrapper", "certifi"]
clus = ["faiss-cpu", "pynndescent", "scikit-learn", "hdbscan", "ray", "joblib", "numba"]

[tool.poetry.dev-dependencies]
black = "^19.10b0"
isort = {extras = ["pyproject"], version = "^4.3.21"}
pre-commit = "^2.4.0"
mypy = "^0.782"
tox = "^3.15.2"
mkdocs = "^1.1.2"
mkdocstrings = "^0.13.6"
mkdocs-material = "^6.1.0"
mkdocs-click = { version = "^0.1.1" }
markdown-inline-graphviz-extension-png = "^1.1"

[build-system]
requires = ["poetry-core>=1.0.0"]
build-backend = "poetry.core.masonry.api"
